SOF-1696 Move mapning of ActionManifest to separate file
