+++ src/blueprints/tv2/action-factories/test/tv2-graphics-action-factory.spec.ts
@@ -10,6 +10,7 @@ import { Tv2StringHashConverter } from '../../helpers/tv2-string-hash-converter'
 import {
   Tv2GraphicsTimelineObjectFactoryFactory
 } from '../../timeline-object-factories/tv2-graphics-timeline-object-factory-factory'
+import { Tv2ActionManifestMapper } from '../../helpers/tv2-action-manifest-mapper'
 
 describe(Tv2GraphicsActionFactory.name, () => {
   it('compiles', () => {
@@ -18,12 +19,14 @@ describe(Tv2GraphicsActionFactory.name, () => {
 })
 
 function createTestee(params?: {
+  actionManifestMapper?: Tv2ActionManifestMapper,
   graphicsTimelineObjectFactoryFactory?: Tv2GraphicsTimelineObjectFactoryFactory,
   audioTimelineObjectFactory?: Tv2AudioTimelineObjectFactory
   videoMixerTimelineObjectFactory?: Tv2VideoMixerTimelineObjectFactory
   stringHashConverter?: Tv2StringHashConverter
 }): Tv2GraphicsActionFactory {
   return new Tv2GraphicsActionFactory(
+    params?.actionManifestMapper ?? instance(mock(Tv2ActionManifestMapper)),
     params?.graphicsTimelineObjectFactoryFactory ?? instance(mock<Tv2GraphicsTimelineObjectFactoryFactory>()),
     params?.audioTimelineObjectFactory ?? instance(mock<Tv2AudioTimelineObjectFactory>()),
     params?.videoMixerTimelineObjectFactory ?? instance(mock<Tv2VideoMixerTimelineObjectFactory>()),
