+++ src/blueprints/tv2/action-factories/test/tv2-video-clip-action-factory.spec.ts
@@ -7,6 +7,7 @@ import {
 } from '../../timeline-object-factories/interfaces/tv2-audio-timeline-object-factory'
 import { Tv2CasparCgTimelineObjectFactory } from '../../timeline-object-factories/tv2-caspar-cg-timeline-object-factory'
 import { instance, mock } from '@typestrong/ts-mockito'
+import { Tv2ActionManifestMapper } from '../../helpers/tv2-action-manifest-mapper'
 
 describe(Tv2VideoClipActionFactory.name, () => {
   it('compiles', () => {
@@ -15,11 +16,13 @@ describe(Tv2VideoClipActionFactory.name, () => {
 })
 
 function createTestee(params?: {
+  actionManifestMapper?: Tv2ActionManifestMapper,
   videoMixerTimelineObjectFactory?: Tv2VideoMixerTimelineObjectFactory,
   audioTimelineObjectFactory?: Tv2AudioTimelineObjectFactory,
   casparCgTimelineObjectFactory?: Tv2CasparCgTimelineObjectFactory
 }): Tv2VideoClipActionFactory {
   return new Tv2VideoClipActionFactory(
+    params?.actionManifestMapper ?? instance(mock(Tv2ActionManifestMapper)),
     params?.videoMixerTimelineObjectFactory ?? instance(mock<Tv2VideoMixerTimelineObjectFactory>()),
     params?.audioTimelineObjectFactory ?? instance(mock<Tv2AudioTimelineObjectFactory>()),
     params?.casparCgTimelineObjectFactory ?? instance(mock(Tv2CasparCgTimelineObjectFactory))
